Begin extracting Incompatible Artifacts describer
